fix: trigger
the npm api/registry and yarn bugged on v1.0.1 version. Even after 30+ minutes after it is released, v1.0.1 still can't be installed.
Installing it locally is okay, but when try to install it inside docker container or using `now` it fails to see it?! wtf?!
Is it related to the https://twitter.com/tunnckoCore/status/1060233900667138049 ?!